MDL-81405 session: New core cookie helper utility class
